requirements.txt
numpy>=1.20
pandas>=1.3
pyarrow>=10.0.0
scikit-learn>=1.0
PyYAML>=6.0

# PennyLane (used when config framework = pennylane)
pennylane>=0.33.0

# Optional: Qiskit (used when config framework = qiskit)
# qiskit>=1.0
# qiskit-machine-learning>=0.7
# qiskit-algorithms>=0.3
